requirements.txt
streamlit==1.24.1
streamlit-folium==0.13.0
pandas==2.0.3
pillow==9.5.0
plotly==5.15.0
numpy==1.24.4
matplotlib==3.7.1
matplotlib-inline==0.1.6
haversine==2.7.0
folium==0.14.0